:hammer: feat: BindingAdapter 수정
- 시간 관련 함수 추가

Issue #78
