+++ presentation/src/main/res/layout/activity_profile.xml
@@ -130,7 +130,7 @@
             app:layout_constraintEnd_toEndOf="parent"
             app:layout_constraintStart_toStartOf="parent"
             app:layout_constraintTop_toBottomOf="@id/container_toolbar"
-            app:visibility="@{vm.isLoading}" />
+            app:toGone="@{!vm.isLoading}" />
 
     </androidx.constraintlayout.widget.ConstraintLayout>
